api/requirements.txt
langchain_groq
langchain_core
langchain-together
langchain_community
youtube-transcript-api
fastapi
uvicorn
python-dotenv